Trim transport session factory indirection
